Create main screen AB#24
