+++ src/app/globals.css
@@ -4,8 +4,8 @@
 
 :root {
   --foreground-rgb: 0, 0, 0;
-  --background-start-rgb: 214, 219, 220;
-  --background-end-rgb: 255, 255, 255;
+  --background-start-rgb: 189, 75, 249;
+  --background-end-rgb: 208, 119, 255;
 }
 
 @media (prefers-color-scheme: dark) {
